Let game determine starting clip size
This replaces the previous CWX behavior of fully manually
replenishing an item's clip to maximum with a call to the game's
functionality to determine the default clip amount.

This allows CWX to gracefully handle the case of weapons that start
with 0 in the clip (the case for weapons that can be overloaded).